glm-4v/requirements.txt
transformers
torch
torchvision
Pillow
accelerate
tiktoken
